pyproject.toml
[tool.poetry]
name = "slack-deleter"
version = "0.1.0"
description = "CLI tools to bulk delete messages in Slack."
authors = ["Theo Diamantidis <diamaltho@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
slackclient = "^2.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
slack-delete = "slack_deleter.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
